+++ src/network_filter_list.rs
@@ -53,7 +53,7 @@ pub enum FlatBufferParsingError {
 
 pub(crate) struct NetworkFilterList {
     pub(crate) memory: VerifiedFlatFilterListMemory,
-    pub(crate) unique_domains_hashes_map: HashMap<Hash, u16>,
+    pub(crate) unique_domains_hashes_map: HashMap<Hash, u32>,
 }
 
 impl Default for NetworkFilterList {
@@ -85,12 +85,12 @@ impl NetworkFilterList {
 
         // Reconstruct the unique_domains_hashes_map from the flatbuffer data
         let len = root.unique_domains_hashes().len();
-        let mut unique_domains_hashes_map: HashMap<crate::utils::Hash, u16> =
+        let mut unique_domains_hashes_map: HashMap<crate::utils::Hash, u32> =
             HashMap::with_capacity(len);
         for (index, hash) in root.unique_domains_hashes().iter().enumerate() {
             unique_domains_hashes_map.insert(
                 hash,
-                u16::try_from(index)
+                u32::try_from(index)
                     .map_err(|_| FlatBufferParsingError::UniqueDomainsOutOfBounds(index))?,
             );
         }
